Move certificate info endpoints to separate service part
